ver0.1

+++ src/main/java/by/platonov/music/repository/GenreRepository.java
@@ -2,8 +2,8 @@ package by.platonov.music.repository;
 
 import by.platonov.music.entity.Genre;
 import by.platonov.music.exception.RepositoryException;
-import by.platonov.music.repository.mapper.AbstractRowMapper;
-import by.platonov.music.repository.mapper.GenreRowMapper;
+import by.platonov.music.repository.mapper.resultSet.AbstractRowMapper;
+import by.platonov.music.repository.mapper.resultSet.GenreRowMapper;
 import by.platonov.music.repository.specification.GenreIdSpecification;
 import by.platonov.music.repository.specification.SqlSpecification;
 import lombok.extern.log4j.Log4j2;
@@ -142,12 +142,12 @@ public class GenreRepository implements Repository<Genre> {
     }
 
     @Override
-    public int count(SqlSpecification specification) throws RepositoryException {
+    public long count(SqlSpecification specification) throws RepositoryException {
         return TransactionHandler.getInstance().transactional(connection -> {
             try (PreparedStatement statement = connection.prepareStatement(SQL_COUNT_GENRE + specification.toSqlClauses());
                  ResultSet resultSet = statement.executeQuery()) {
                 resultSet.next();
-                return resultSet.getInt(1);
+                return resultSet.getLong(1);
             } catch (SQLException e) {
                 throw new RepositoryException(e);
             }
