+++ src/test/java/by/platonov/music/repository/MusicianRepositoryTest.java
@@ -47,8 +47,8 @@ class MusicianRepositoryTest {
         repository.add(newMusician);
 
         //then
-        int actual = repository.count(new IdIsNotNullSpecification());
-        int expected = 8;
+        long actual = repository.count(new IdIsNotNullSpecification());
+        long expected = 8;
         assertEquals(expected, actual);
     }
 
@@ -58,8 +58,8 @@ class MusicianRepositoryTest {
         repository.add(existsMusician);
 
         //then
-        int actual = repository.count(new IdIsNotNullSpecification());
-        int expected = 7;
+        long actual = repository.count(new IdIsNotNullSpecification());
+        long expected = 7;
         assertEquals(expected, actual);
     }
 
@@ -93,8 +93,8 @@ class MusicianRepositoryTest {
         repository.remove(existsMusician);
 
         //then
-        int actual = repository.count(new IdIsNotNullSpecification());
-        int expected = 6;
+        long actual = repository.count(new IdIsNotNullSpecification());
+        long expected = 6;
         assertEquals(expected, actual);
     }
 
@@ -104,8 +104,8 @@ class MusicianRepositoryTest {
         repository.remove(newMusician);
 
         //then
-        int actual = repository.count(new IdIsNotNullSpecification());
-        int expected = 7;
+        long actual = repository.count(new IdIsNotNullSpecification());
+        long expected = 7;
         assertEquals(expected, actual);
     }
 
@@ -135,8 +135,8 @@ class MusicianRepositoryTest {
 
     @Test
     void count() throws RepositoryException {
-        int actual = repository.count(new IdIsNotNullSpecification());
-        int expected = 7;
+        long actual = repository.count(new IdIsNotNullSpecification());
+        long expected = 7;
         assertEquals(expected, actual);
     }
 }
